pyproject.toml
# pyproject.toml

[project]
name = "DLOlympus"
authors = [
    {name = "Facundo Roffet", email = "facundo.roffet@cs.uns.edu.ar"},
]
version = "0.2.0"
description = "Tools and pipelines for Deep Learning."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "albumentations>=2.0.8",
    "dill>=0.4.0",
    "fastai>=2.7.18",
    "fastcore>=1.7.29",
    "matplotlib>=3.10.6",
    "nbformat>=5.10.4",
    "numpy>=2.3.2",
    "onnx>=1.19.0",
    "onnxruntime>=1.22.1",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.2",
    "Pillow>=11.3.0",
    "plum-dispatch>=2.5.7",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "timm>=1.0.19",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "wandb>=0.21.1",
]

[project.optional-dependencies]
dev = [
    "hypothesis",
    "pytest",
    "pipreqs",
    "pytest-mock",
    "ruff",
]